extract response code modifying resource provider for testing
