+++ hapi-fhir-structures-r4/src/test/java/ca/uhn/fhir/rest/server/ResponseCodeModifyingResourceProviderTest.java
@@ -1,32 +1,23 @@
 package ca.uhn.fhir.rest.server;
 
 import ca.uhn.fhir.context.FhirVersionEnum;
-import ca.uhn.fhir.rest.annotation.Create;
-import ca.uhn.fhir.rest.annotation.IdParam;
-import ca.uhn.fhir.rest.annotation.ResourceParam;
-import ca.uhn.fhir.rest.annotation.Update;
+import ca.uhn.fhir.provider.TestResponseCodeModifyingPatientProvider;
 import ca.uhn.fhir.rest.api.MethodOutcome;
 import ca.uhn.fhir.rest.client.api.IGenericClient;
 import ca.uhn.fhir.test.utilities.server.RestfulServerExtension;
-import org.hl7.fhir.instance.model.api.IBaseResource;
-import org.hl7.fhir.r4.model.IdType;
 import org.hl7.fhir.r4.model.Patient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-import javax.servlet.ServletRequest;
-
-import static ca.uhn.fhir.rest.api.Constants.STATUS_HTTP_202_ACCEPTED;
+import static ca.uhn.fhir.provider.TestResponseCodeModifyingPatientProvider.CUSTOM_RESPONSE_CODE;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class ResponseCodeModifyingResourceProviderTest {
 
-	public static final int CUSTOM_RESPONSE_CODE = STATUS_HTTP_202_ACCEPTED;
-
 	@RegisterExtension
 	private RestfulServerExtension myServer = new RestfulServerExtension(FhirVersionEnum.R4)
-		 .registerProvider(new ResponseCodeModifyingPatientProvider());
+		 .registerProvider(new TestResponseCodeModifyingPatientProvider());
 
 	private IGenericClient myClient;
 
@@ -50,28 +41,6 @@ public class ResponseCodeModifyingResourceProviderTest {
 		myPatient.setId("1");
 		MethodOutcome outcome = myClient.update().resource(myPatient).execute();
 		assertEquals(CUSTOM_RESPONSE_CODE, outcome.getResponseStatusCode());
-	}
-
-	public static class ResponseCodeModifyingPatientProvider implements IResourceProvider {
-
-		@Create()
-		public MethodOutcome createPatient(@ResourceParam Patient thePatient, ServletRequest theServletRequest) {
-			MethodOutcome methodOutcome = new MethodOutcome();
-			methodOutcome.setResponseStatusCode(CUSTOM_RESPONSE_CODE);
-			return methodOutcome;
-		}
-
-		@Update()
-		public MethodOutcome updatePatient(@IdParam IdType theId, @ResourceParam Patient thePatient) {
-			MethodOutcome methodOutcome = new MethodOutcome();
-			methodOutcome.setResponseStatusCode(CUSTOM_RESPONSE_CODE);
-			return methodOutcome;
-		}
-
-		@Override
-		public Class<? extends IBaseResource> getResourceType() {
-			return Patient.class;
-		}
 	}
 
 }
